- fixed path deployments - fixed already defined apps redirection - simplified prechecks before deployment
