+++ api/routes/v1/application/deploy/index.js
@@ -1,8 +1,8 @@
 
-const { verifyUserId, cleanupTmp, execShellAsync } = require('../../../../libs/common')
+const { verifyUserId, cleanupTmp } = require('../../../../libs/common')
 const Deployment = require('../../../../models/Deployment')
 const { queueAndBuild } = require('../../../../libs/applications')
-const { setDefaultConfiguration } = require('../../../../libs/applications/configuration')
+const { setDefaultConfiguration, precheckDeployment } = require('../../../../libs/applications/configuration')
 const { docker } = require('../../../../libs/docker')
 const cloneRepository = require('../../../../libs/applications/github/cloneRepository')
 
@@ -39,67 +39,15 @@ module.exports = async function (fastify) {
       return
     }
     try {
-      const configuration = setDefaultConfiguration(request.body)
-
       const services = (await docker.engine.listServices()).filter(r => r.Spec.Labels.managedBy === 'coolify' && r.Spec.Labels.type === 'application')
-
+      const configuration = setDefaultConfiguration(request.body)
       await cloneRepository(configuration)
+      const { foundService, imageChanged, configChanged } = await precheckDeployment({ services, configuration })
 
-      let foundService = false
-      let foundDomain = false
-      let configChanged = false
-      let imageChanged = false
-
-      let forceUpdate = false
-
-      for (const service of services) {
-        const running = JSON.parse(service.Spec.Labels.configuration)
-        if (running) {
-          if (
-            running.publish.domain === configuration.publish.domain &&
-            running.repository.id !== configuration.repository.id
-          ) {
-            foundDomain = true
-          }
-          if (running.repository.id === configuration.repository.id && running.repository.branch === configuration.repository.branch) {
-            // Base service configuration changed
-            if (!running.build.container.baseSHA || running.build.container.baseSHA !== configuration.build.container.baseSHA) {
-              configChanged = true
-            }
-            const state = await execShellAsync(`docker stack ps ${running.build.container.name} --format '{{ json . }}'`)
-            const isError = state.split('\n').filter(n => n).map(s => JSON.parse(s)).filter(n => n.DesiredState !== 'Running')
-            if (isError.length > 0) forceUpdate = true
-
-            foundService = true
-            const runningWithoutContainer = JSON.parse(JSON.stringify(running))
-            delete runningWithoutContainer.build.container
-
-            const configurationWithoutContainer = JSON.parse(JSON.stringify(configuration))
-            delete configurationWithoutContainer.build.container
-
-            // If only the configuration changed
-            if (JSON.stringify(runningWithoutContainer.build) !== JSON.stringify(configurationWithoutContainer.build) || JSON.stringify(runningWithoutContainer.publish) !== JSON.stringify(configurationWithoutContainer.publish)) configChanged = true
-            // If only the image changed
-            if (running.build.container.tag !== configuration.build.container.tag) imageChanged = true
-            // If build pack changed, forceUpdate the service
-            if (running.build.pack !== configuration.build.pack) forceUpdate = true
-          }
-        }
-      }
-      if (foundDomain) {
+      if (foundService && !imageChanged && !configChanged) {
         cleanupTmp(configuration.general.workdir)
-        reply.code(500).send({ message: 'Domain already in use.' })
+        reply.code(500).send({ message: 'Nothing changed, no need to redeploy.' })
         return
-      }
-      if (forceUpdate) {
-        imageChanged = false
-        configChanged = false
-      } else {
-        if (foundService && !imageChanged && !configChanged) {
-          cleanupTmp(configuration.general.workdir)
-          reply.code(500).send({ message: 'Nothing changed, no need to redeploy.' })
-          return
-        }
       }
 
       const alreadyQueued = await Deployment.find({
