+++ api/routes/v1/webhooks/deploy.js
@@ -1,8 +1,8 @@
 const crypto = require('crypto')
-const { cleanupTmp, execShellAsync } = require('../../../libs/common')
+const { cleanupTmp } = require('../../../libs/common')
 const Deployment = require('../../../models/Deployment')
 const { queueAndBuild } = require('../../../libs/applications')
-const { setDefaultConfiguration } = require('../../../libs/applications/configuration')
+const { setDefaultConfiguration, precheckDeployment } = require('../../../libs/applications/configuration')
 const { docker } = require('../../../libs/docker')
 const cloneRepository = require('../../../libs/applications/github/cloneRepository')
 
@@ -68,59 +68,14 @@ module.exports = async function (fastify) {
       }
 
       configuration = setDefaultConfiguration(JSON.parse(configuration.Spec.Labels.configuration))
-
       await cloneRepository(configuration)
+      const { foundService, imageChanged, configChanged } = await precheckDeployment({ services, configuration })
 
-      let foundService = false
-      let foundDomain = false
-      let configChanged = false
-      let imageChanged = false
-
-      let forceUpdate = false
-
-      for (const service of services) {
-        const running = JSON.parse(service.Spec.Labels.configuration)
-        if (running) {
-          if (
-            running.publish.domain === configuration.publish.domain &&
-            running.repository.id !== configuration.repository.id &&
-            running.repository.branch !== configuration.repository.branch
-          ) {
-            foundDomain = true
-          }
-          if (running.repository.id === configuration.repository.id && running.repository.branch === configuration.repository.branch) {
-            const state = await execShellAsync(`docker stack ps ${running.build.container.name} --format '{{ json . }}'`)
-            const isError = state.split('\n').filter(n => n).map(s => JSON.parse(s)).filter(n => n.DesiredState !== 'Running')
-            if (isError.length > 0) forceUpdate = true
-            foundService = true
-
-            const runningWithoutContainer = JSON.parse(JSON.stringify(running))
-            delete runningWithoutContainer.build.container
-
-            const configurationWithoutContainer = JSON.parse(JSON.stringify(configuration))
-            delete configurationWithoutContainer.build.container
-
-            if (JSON.stringify(runningWithoutContainer.build) !== JSON.stringify(configurationWithoutContainer.build) || JSON.stringify(runningWithoutContainer.publish) !== JSON.stringify(configurationWithoutContainer.publish)) configChanged = true
-            if (running.build.container.tag !== configuration.build.container.tag) imageChanged = true
-          }
-        }
-      }
-      if (foundDomain) {
+      if (foundService && !imageChanged && !configChanged) {
         cleanupTmp(configuration.general.workdir)
-        reply.code(500).send({ message: 'Domain already used.' })
+        reply.code(500).send({ message: 'Nothing changed, no need to redeploy.' })
         return
       }
-      if (forceUpdate) {
-        imageChanged = false
-        configChanged = false
-      } else {
-        if (foundService && !imageChanged && !configChanged) {
-          cleanupTmp(configuration.general.workdir)
-          reply.code(500).send({ message: 'Nothing changed, no need to redeploy.' })
-          return
-        }
-      }
-
       const alreadyQueued = await Deployment.find({
         repoId: configuration.repository.id,
         branch: configuration.repository.branch,
@@ -135,9 +90,9 @@ module.exports = async function (fastify) {
         return
       }
 
-      queueAndBuild(configuration, services, configChanged, imageChanged)
+      queueAndBuild(configuration, configChanged, imageChanged)
 
-      reply.code(201).send({ message: 'Deployment queued.' })
+      reply.code(201).send({ message: 'Deployment queued.', nickname: configuration.general.nickname, name: configuration.build.container.name })
     } catch (error) {
       throw { error, type: 'server' }
     }
