+++ src/components/Application/Configuration/Configuration.svelte
@@ -29,7 +29,7 @@
 
   async function loadBranches() {
     loading.branches = true;
-    if ($isActive("/application/new")) $application.repository.branch = null
+    if ($isActive("/application/new")) $application.repository.branch = null;
     const selectedRepository = repositories.find(
       r => r.id === $application.repository.id,
     );
@@ -54,6 +54,7 @@
   }
 
   async function loadGithub() {
+    loading.github = true;
     try {
       const { installations } = await $fetch(
         "https://api.github.com/user/installations",
@@ -100,7 +101,6 @@
     } finally {
       loading.github = false;
     }
- 
   }
   function modifyGithubAppConfig() {
     const left = screen.width / 2 - 1020 / 2;
@@ -143,6 +143,52 @@
     }, 100);
   }
 </script>
+
+{#if !$isActive("/application/new")}
+  <div class="min-h-full text-white">
+    <div
+      class="py-5 text-left px-6 text-3xl tracking-tight font-bold flex items-center"
+    >
+      <a
+        target="_blank"
+        class="text-green-500 hover:underline cursor-pointer px-2"
+        href="{'https://' +
+          $application.publish.domain +
+          $application.publish.path}"
+        >{$application.publish.domain
+          ? $application.publish.domain + $application.publish.path
+          : "Loading..."}</a
+      >
+      <a
+        target="_blank"
+        class="icon"
+        href="{`https://github.com/${$application.repository.organization}/${$application.repository.name}`}"
+      >
+        <svg
+          class="w-6"
+          xmlns="http://www.w3.org/2000/svg"
+          viewBox="0 0 24 24"
+          fill="none"
+          stroke="currentColor"
+          stroke-width="2"
+          stroke-linecap="round"
+          stroke-linejoin="round"
+          ><path
+            d="M9 19c-5 1.5-5-2.5-7-3m14 6v-3.87a3.37 3.37 0 0 0-.94-2.61c3.14-.35 6.44-1.54 6.44-7A5.44 5.44 0 0 0 20 4.77 5.07 5.07 0 0 0 19.91 1S18.73.65 16 2.48a13.38 13.38 0 0 0-7 0C6.27.65 5.09 1 5.09 1A5.07 5.07 0 0 0 5 4.77a5.44 5.44 0 0 0-1.5 3.78c0 5.42 3.3 6.61 6.44 7A3.37 3.37 0 0 0 9 18.13V22"
+          ></path></svg
+        ></a
+      >
+    </div>
+  </div>
+{:else if $isActive("/application/new")}
+  <div class="min-h-full text-white">
+    <div
+      class="py-5 text-left px-6 text-3xl tracking-tight font-bold flex items-center"
+    >
+      New Application
+    </div>
+  </div>
+{/if}
 
 <div in:fade="{{ duration: 100 }}">
   {#if !$session.githubAppToken}
