+++ src/pages/application/new.svelte
@@ -2,12 +2,4 @@
   import Configuration from "../../components/Application/Configuration/Configuration.svelte";
 </script>
 
-<div class="min-h-full text-white">
-  <div
-    class="py-5 text-left px-6 text-3xl tracking-tight font-bold flex items-center"
-  >
-    New Application
-  </div>
-</div>
-
 <Configuration />
